my.kython.konsume: move to core
